further clean up, added rt interval
git-svn-id: https://svn.mi.fu-berlin.de/des-testbed/Software/DES-SERT/des-olsr/trunk@7710 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ src/database/rl_seq_t/rl_seq.h
@@ -34,7 +34,7 @@ u_int16_t rl_get_nextseq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALE
  * Return TRUE if given seq_num is equals to thin in database
  * (i.e. this packet was already processed)
  */
-int rl_check_seq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN], u_int16_t seq_num);
+uint8_t rl_check_seq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN], u_int16_t seq_num);
 
 void rl_add_seq(u_int8_t src_addr[ETH_ALEN], u_int8_t dest_addr[ETH_ALEN], u_int16_t seq_num);
